requirements.txt
apache-airflow==3.1.0
apache-airflow-providers-slack==9.3.0
pandas==2.3.3
requests==2.32.5
pyarrow==21.0.0
pydantic-settings==2.11.0
awscli==1.42.59
boto3==1.40.59




